オブジェクト毎に分割し，岩を各レーンで表示可能にした. 0,1,2 を-1,0,1にした
