+++ View.java
@@ -59,121 +59,11 @@ public class View extends JFrame {
         setVisible(true);
     }
     
-    /////////////////////////////////////////////////////////////////
-///背景クラス
-    public class LanePanel extends JPanel {
-        
-        @Override
-        public void paintComponent(Graphics g) {
-            super.paintComponent(g);
-            Graphics2D g2d = (Graphics2D) g;
-            Color DarkGreen = new Color(0, 100, 0);
-            
-            // グラデーションの設定
-            GradientPaint gradient = new GradientPaint(
-                0, 0, Color.GREEN, // 開始位置と色
-                0, getHeight(), DarkGreen// 終了位置と色
-                );
-    
-                // グラデーションを適用して塗りつぶす
-                g2d.setPaint(gradient);
-                g2d.fillRect(0, 0, getWidth(), getHeight());
-            }
-        }
-        
-//////////////////////////////////////////////////////////////////////////
-///プレーヤークラス
-public class PlayerPanel extends JPanel{
-    // メンバ変数
-    int x, y, width, height; // x=0, 1, 2で位置決定
-    Image image;
-    //private int playerPosX = 1;
-    
-    // コンストラクタ
-    public PlayerPanel(){
-            x = 1; y = 0;
-            width = 100; height = 100;
-            try{
-                image = ImageIO.read(new File("URL"));
-            }catch(IOException e){
-                e.printStackTrace();
-            }
-            setOpaque(false);//背景を透過するやつ。
-        }
-        
-        public void paintComponent(Graphics g){
-            int offsetX = 50, offsetY = 700;
-            super.paintComponent(g);
-            if(image!=null){
-                g.drawImage(image, x*200+offsetX, y*100+offsetY, getFocusCycleRootAncestor());
-            }else{
-                g.setColor(Color.CYAN);
-                g.fillOval(x*200+offsetX, y*100+offsetY, width, height);
-            }
-        }
-        
-        // プレイヤー位置を更新して再描画.
-        //★★Controllerで呼ばれる。
-        public void updatePlayerPos(int playerPosX) {
-            this.x = playerPosX;
-            repaint();
-        }
-    }
-    /////////////////////////////////////////////////////////////////////////
-    /// 岩クラス
-    public class RockPanel extends JPanel{
-        // メンバ変数
-        int x, y, width, height; // x=0, 1, 2で位置決定
-        Image image;
-        //private int playerPosX = 1;
-        
-        // コンストラクタ
-        public RockPanel(){
-            x = 1; y = 0;
-            width = 100; height = 100;
-            try{
-                image = ImageIO.read(new File("URL"));
-            }catch(IOException e){
-                e.printStackTrace();
-            }
-            setOpaque(false);//背景を透過するやつ。
-        }
-        
-        public void paintComponent(Graphics g){
-            int offsetX = 50, offsetY = 200;
-            super.paintComponent(g);
-            if(image!=null){
-                g.drawImage(image, x*200+offsetX, y*100+offsetY, getFocusCycleRootAncestor());
-            }else{
-                g.setColor(Color.ORANGE);
-                g.fillOval(x*200+offsetX, (int)(y*1.35+offsetY), width, height);
-
-                // はまぶーによる変更
-                //g.fillOval(x*200+offsetX, y*100+offsetY, width, height);
-            }
-        }
-        
-        // プレイヤー位置を更新して再描画.
-        //★★Controllerで呼ばれる。
-        public void updateRockPos(int rockPosY) {
-            this.y = rockPosY;
-            repaint();
-        }
-
-        // はまぶーによる変更
-
-        // public void updateRockPos(int playerPosX) {
-        //     this.x = playerPosX;
-        //     repaint();
-        // }
-    }
-
-    public PlayerPanel getPlayerPanel() {
-        return playerPanel;
-    }
-
     public RockPanel getRockPanel() {
         return rockPanel;
     }
 
+        public PlayerPanel getPlayerPanel() {
+        return playerPanel;
+    }
 }
